Check fifo file with the S_ISFIFO macro
This commit contains two changes:

Use the `S_ISFIFO` macro to hide the bitwise operations and prevent the
module from being affected by the kernel in case the kernel changes the
name of `S_IFIFO` or `S_IFMT`.

And aside from using the macro, this commit also reduce the use the
inode pointer which is not used in other places.